Get Data on EU Infringement Cases

This function provides an easy way to get case-level data from the European Union Infringement Procedures (EUIP) dataset API without writing an API query and parsing the JSON data returned by the API. This function returns one observation per infringement case that matches all of the search criteria you specify.

get_case_data(
  id = NULL,
  year = NULL,
  ms = NULL,
  dg = NULL,
  dir = NULL,
  celex = NULL,
  type = NULL,
  stage = NULL,
  comp = FALSE
)

Arguments

id
A numeric vector where each number is a valid case number. Case numbers always have 8 digits. The first 4 digits are the year the case was opened and the second 4 digits uniquely identify the cases within the year. If you include one or more invalid case numbers, you will get an error.

year
A 4-digit numeric value indicating a year or a numeric vector of length 2 containing a start year and an end year. If you list more than 2 years, or do not list the start year and end year in chronological order, you will get an error. The EUIP dataset covers the period from 2002 to 2020. If you include a year outside this range, you will also get an error.

ms
A string vector where each string is a valid member state name or code. Use the function ms_list() to get a complete list of valid member state names and codes. The API will return an error if one or more of the elements is not valid.

dg
A string vector where each string is a valid DG name or code. Use the function dg_list() to get a complete list of valid DG names and codes. The API will return an error if one or more of the elements is not valid.

dir
A vector of valid directive numbers. Note that not all observations correspond to a specific directive. If a observation does not, and this parameter is not NULL, the API will not return it, even if it meets all of the other specified criteria. All observations related to noncommunication cases should correspond to a specific directive, but the Commission does not always report this information. If an invalid directive number is included, the API will return an error.

celex
A vector of valid CELEX numbers for directives. Note that not all observations correspond to a specific directive. If a observation does not, and this parameter is not NULL, the API will not return it, even if it meets all of the other specified criteria. All observations related to noncommunication cases should correspond to a specific directive, but the Commission does not always report this information. If an invalid CELEX number is included, the API will return an error.

type
A string vector where each string is a type of infringment case. Valid options include:

"COMM", for noncommunication cases; and

"CONF", for nonconformity cases.

If NULL, the function will return both types of cases. Noncommunication cases are those in which the Commission alleges that a member state has failed to notify the Commission of the transposition of a directive before the specified date. Nonconformity cases are those in which the Commission alleges that a member state has failed to correctly transpose a directive into national law.

stage
A string vector where each string is a stage in the infringement procedure. Only cases that make it to all specified stages will be included in the results. Valid options include:

"LFN258", for letters of formal notice under Article 258 (TFEU);

"RO258", for reasoned opinions under Article 258 (TFEU);

"RF258", for referrals to the Court under Article 258 (TFEU);

"LFN260", for letters of formal notice under Article 260 (TFEU);

"RO260", for reasoned opinions under Article 260 (TFEU);

"RF260", for referrals to the Court under Article 260 (TFEU);

"C", for the closing of the case; and

"W", for the withdrawal of the case.

Note that a case can be closed or withdrawn, but not both.

comp
A logical value (TRUE or FALSE). If TRUE, only complete cases (cases with no known missing stages) are included in the results. For example, a case with a letter of formal notice and a referral to the Court, but no recorded reasoned opinion would be omitted from the results. To not be omitted, a case must at least have a record of a letter of formal notice and a record of a closing or a withdrawal.

Value

Returns a tibble in tidy format with case-level data, subject to the specified filters. The unit of observation is an infringement case. The tibble will include the following variables:

case_number: a unique ID number assigned to the case by the Commission;

case_year: the year the case was opened (i.e., the date of the letter of formal notice under Article 258 TFEU);

member_state: the name of the member state under investigation;

member_state_code: the code of the member state under investigation;

directorate_general: the name of the DG that managed the case;

directorate_general_code: the code of the DG that managed the case;

case_type: either "Noncommunication" or "Nonconformity";

stage_LFN258: a dummy variable indicating whether there was an letter of formal notice under Article 258 (TFEU);

stage_RO258: a dummy variable indicating whether there was a reasoned opinion under Article 258 (TFEU);

stage_RF258: a dummy variable indicating whether there was a referal to the Court under Article 258 (TFEU);

stage_LFN260: a dummy variable indicating whether there was an letter of formal notice under Article 260 (TFEU);

stage_RO260: a dummy variable indicating whether there was a reasoned opinion under Article 260 (TFEU);

stage_RF260: a dummy variable indicating whether there was a referal to the Court under Article 260 (TFEU);

stage_closing: a dummy variable indicating whether there is a record of the case being closed; and

stage_withdrawal: a dummy variable indicating whether there is a record of the case being withdrawn.

Details

The EUIP dataset API returns data in a standard JSON format. This function automatically converts the JSON data into a tibble (see the tidyverse package). The output of this function will be in a tidy format (each variable in its own column, each observation in its own row, and each value in its own cell) for easy use with other functions in the tidyverse, such as ggplot.

See also