Get Directorate-General Data

This function provides an easy way to get time-varying DG-level data from the European Union Infringement Procedures (EUIP) dataset API without writing an API query and parsing the JSON data returned by the API. This function returns time-varying DG-level data on the number decisions made by the Commission in infringement cases at each stage of the infringement procedure (letter of formal notice, reasoned opinion, etc.) with an option to disaggregate by type of case (e.g., noncommunication vs nonconformity). Use the arguments to apply filters that narrow down the results to data for specific years, specific DGs, specific types of cases, and/or specific stages of the infringement procedure.

get_dg_data(year = NULL, dg = NULL, type = NULL, stage = NULL)

Arguments

year
A four-digit numeric value indicating a year or a numeric vector of length 2 containing a start year and an end year. If you list more than 2 years, or do not list the start year and end year in chronological order, you will get an error. The EUIP dataset covers the period from 2002 to 2020. If you include a year outside this range, you will also get an error.

dg
A string vector where each string is a valid DG name or code. Use the function dg_list() to get a complete list of valid DG names and codes. The API will return an error if one or more of the elements is not valid.

type
A string vector where each string is a type of infringment case. Valid options include:

"COMM", for noncommunication cases; and

"CONF", for nonconformity cases.

If NULL, the function will not distinguish between noncommunication cases and nonconformity cases. Noncommunication cases are those in which the Commission alleges that a member state has failed to notify the Commission of the transposition of a directive before the specified date. Nonconformity cases are those in which the Commission alleges that a member state has failed to correctly transpose a directive into national law.

stage
A string vector where each string is a stage in the infringement procedure. Valid options include:

"LFN258", for letters of formal notice under Article 258 (TFEU);

"RO258", for reasoned opinions under Article 258 (TFEU);

"RF258", for referrals to the Court under Article 258 (TFEU);

"LFN260", for letters of formal notice under Article 260 (TFEU);

"RO260", for reasoned opinions under Article 260 (TFEU); and

"RF260", for referrals to the Court under Article 260 (TFEU).

Value

Returns a tibble in tidy format with time-varying DG-level data on the number decisions made by the Commission in infringement cases at each stage of the infringement procedure (letter of formal notice, reasoned opinion, etc.) and by type of case (e.g., noncommunication vs nonconformity), subject to the specified filters. The unit of observation is the unique combination of a year, a DG, a type of case, and a stage of the infringement procedure. The tibble will include the following variables:

directorate_general: the name of the member state;

directorate_general_code: a 2-character code for the member state;

year: the year of the decision (not the year the case was lodged);

decision_stage: the stage of the infringement procedure;

case_type: if type is not NULL, the type of case ("Noncommunication" or "Nonconformity"), otherwise excluded; and

count: the number of decisions in cases of the indicated type at the indicated stage involving the indicated DG in the indicated year.

Details

The EUIP dataset API returns data in a standard JSON format. This function automatically converts the JSON data into a tibble (see the tidyverse package). The output of this function is in a tidy format (each variable in its own column, each observation in its own row, and each value in its own cell) for easy use with other functions in the tidyverse, such as ggplot.

See also

Examples

# the number of letters of formal notice under Article 258
# sent by DG ENV and DG ENER in noncommunication cases
# for each year from 2010 through 2010

get_dg_data(year = c(2010, 2020), dg = c("ENV", "ENER"), type = "COMM", stage = "LFN258")

#> # A tibble: 22 x 6
#>    directorate_gene… directorate_gener…  year decision_stage    case_type  count
#>    <chr>             <chr>              <int> <chr>             <chr>      <int>
#>  1 Energy            ENER                2010 Letter of formal… Noncommun…     0
#>  2 Environment       ENV                 2010 Letter of formal… Noncommun…   169
#>  3 Energy            ENER                2011 Letter of formal… Noncommun…   140
#>  4 Environment       ENV                 2011 Letter of formal… Noncommun…    58
#>  5 Energy            ENER                2012 Letter of formal… Noncommun…    24
#>  6 Environment       ENV                 2012 Letter of formal… Noncommun…    63
#>  7 Energy            ENER                2013 Letter of formal… Noncommun…    33
#>  8 Environment       ENV                 2013 Letter of formal… Noncommun…   168
#>  9 Energy            ENER                2014 Letter of formal… Noncommun…    25
#> 10 Environment       ENV                 2014 Letter of formal… Noncommun…   127
#> # … with 12 more rows

# the number of decisions that the Commission made in cases
# managed by DG GROW by type in nonconformity cases in 2010

get_dg_data(year = 2019, dg = "GROW", type = "CONF")

#> # A tibble: 6 x 6
#>   directorate_general   directorate_gene…  year decision_stage   case_type count
#>   <chr>                 <chr>             <int> <chr>            <chr>     <int>
#> 1 Internal Market, Ind… GROW               2019 Letter of forma… Nonconfo…   127
#> 2 Internal Market, Ind… GROW               2019 Reasoned opinio… Nonconfo…    55
#> 3 Internal Market, Ind… GROW               2019 Referral to the… Nonconfo…     2
#> 4 Internal Market, Ind… GROW               2019 Letter of forma… Nonconfo…     2
#> 5 Internal Market, Ind… GROW               2019 Reasoned opinio… Nonconfo…     0
#> 6 Internal Market, Ind… GROW               2019 Referral to the… Nonconfo…     0