Get Member State Data

This function provides an easy way to get time-varying member state-level data from the European Union Infringement Procedures (EUIP) dataset API without writing an API query and parsing the JSON data returned by the API. This function returns time-varying member state-level data on the number decisions made by the Commission in infringement cases at each stage of the infringement procedure (letter of formal notice, reasoned opinion, etc.) with an option to disaggregate by type of case (noncommunication vs nonconformity). Use the arguments to apply filters that narrow down the results to data for specific years, specific member states, specific types of cases, and/or specific stages of the infringement procedure.

get_ms_data(year = NULL, ms = NULL, type = NULL, stage = NULL)

Arguments

year
A 4-digit numeric value indicating a year or a numeric vector of length 2 containing a start year and an end year. If you list more than 2 years, or do not list the start year and end year in chronological order, you will get an error. The EUIP dataset covers the period from 2002 to 2020. If you include a year outside this range, you will also get an error.

ms
A string vector where each string is a valid member state name or code. Use the function ms_list() to get a complete list of valid member state names and codes. The API will return an error if one or more of the elements is not valid.

type
A string vector where each string is a type of infringment case. Valid options include:

"COMM", for noncommunication cases; and

"CONF", for nonconformity cases.

If NULL, the function will not distinguish between noncommunication cases and nonconformity cases. Noncommunication cases are those in which the Commission alleges that a member state has failed to notify the Commission of the transposition of a directive before the specified date. Nonconformity cases are those in which the Commission alleges that a member state has failed to correctly transpose a directive into national law.

stage
A string vector where each string is a stage in the infringement procedure. Valid options include:

"LFN258", for letters of formal notice under Article 258 (TFEU);

"RO258", for reasoned opinions under Article 258 (TFEU);

"RF258", for referrals to the Court under Article 258 (TFEU);

"LFN260", for letters of formal notice under Article 260 (TFEU);

"RO260", for reasoned opinions under Article 260 (TFEU); and

"RF260", for referrals to the Court under Article 260 (TFEU).

Value

Returns a tibble in tidy format with time-varying member state-level data on the number decisions made by the Commission in infringement cases at each stage of the infringement procedure (letter of formal notice, reasoned opinion, etc.) and by type of case (e.g., noncommunication vs nonconformity), subject to the specified filters. The unit of observation is the unique combination of a member state, a year, a type of case, and a stage of the infringement procedure. The tibble will include the following variables:

member_state: the name of the member state;

member_state_code: a 2-character code for the member state;

year: the year of the decision (not the year the case was lodged);

decision_stage: the stage of the infringement procedure;

case_type: if type is not NULL, the type of case ("Noncommunication" or "Nonconformity"), otherwise excluded; and

count: a count of decisions.

Details

The EUIP dataset API returns data in a standard JSON format. This function automatically converts the JSON data into a tibble (see the tidyverse package). The output of this function is in a tidy format (each variable in its own column, each observation in its own row, and each value in its own cell) for easy use with other functions in the tidyverse, such as ggplot.

See also

Examples

# the number of letters of formal notice under Article 258
# that the Commission sent to the United Kingdom,
# France, and Germany in noncommunication cases
# for each year from 2010 through 2010

get_ms_data(year = c(2010, 2020), ms = c("UK", "FR", "DE"), type = "COMM", stage = "LFN258")

#> # A tibble: 33 x 6
#>    member_state  member_state_code  year decision_stage        case_type   count
#>    <chr>         <chr>             <int> <chr>                 <chr>       <int>
#>  1 France        FR                 2010 Letter of formal not… Noncommuni…    28
#>  2 Germany       DE                 2010 Letter of formal not… Noncommuni…    21
#>  3 United Kingd… UK                 2010 Letter of formal not… Noncommuni…    35
#>  4 France        FR                 2011 Letter of formal not… Noncommuni…    43
#>  5 Germany       DE                 2011 Letter of formal not… Noncommuni…    32
#>  6 United Kingd… UK                 2011 Letter of formal not… Noncommuni…    58
#>  7 France        FR                 2012 Letter of formal not… Noncommuni…    16
#>  8 Germany       DE                 2012 Letter of formal not… Noncommuni…    12
#>  9 United Kingd… UK                 2012 Letter of formal not… Noncommuni…    24
#> 10 France        FR                 2013 Letter of formal not… Noncommuni…    15
#> # … with 23 more rows

# the number of decisions that the Commission made in cases
# against France by type in nonconformity cases in 2010

get_ms_data(year = 2010, ms = "FR", type = "CONF")

#> # A tibble: 6 x 6
#>   member_state member_state_code  year decision_stage           case_type  count
#>   <chr>        <chr>             <int> <chr>                    <chr>      <int>
#> 1 France       FR                 2010 Letter of formal notice… Nonconfor…    16
#> 2 France       FR                 2010 Reasoned opinion (Art. … Nonconfor…    15
#> 3 France       FR                 2010 Referral to the Court (… Nonconfor…     9
#> 4 France       FR                 2010 Letter of formal notice… Nonconfor…     1
#> 5 France       FR                 2010 Reasoned opinion (Art. … Nonconfor…     0
#> 6 France       FR                 2010 Referral to the Court (… Nonconfor…     0